Move styled-react hooks out of component modules
